+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * fix(examples): fix `LV_USE_DEMO_WIDGETS` typo by @lboue (#98)
 * fix(examples): fix `Tearing fucntion` typo by @lboue (#96)
 * fix(examples): fix WiFiClock log HTTP error code to serial console by @lboue (#97)
+* fix(examples): fix WiFiClock description
 * fix(gt911): allow to set the GT911 touch device address by @lboue (#86)
 * fix(conf): fix the issue that the `ESP_PANEL_EXPANDER_HOST_ID` flag is not working properly
 * fix(conf): fix `LCD Venbdor` typo (#92)